git-svn-id: https://svn.neo4j.org/components@2845 0b971d98-bb2f-0410-8247-b05b2b5feb2a

+++ neo/branches/b8-unsigned/src/java/org/neo4j/impl/nioneo/store/PersistenceWindowPool.java
@@ -137,7 +137,7 @@ class PersistenceWindowPool
                 miss++;
                 brickMiss++;
                 
-                PersistenceRow dpw = activeRowWindows.get( position ); 
+                PersistenceRow dpw = activeRowWindows.get( (int) position ); 
                 
                 if ( dpw == null )
                 {
